requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.0
decorator==5.1.1
Django==5.1.5
django-allauth==65.3.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-markdownx==4.0.7
executing==2.1.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
Markdown==3.7
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
